# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "tech-software-engineer-consultant-int-backend, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Facture/Create-Facture": {
      "post": {
        "tags": [
          "Facture"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FactureCreateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Facture/get-Facture-ID/{id}": {
      "get": {
        "tags": [
          "Facture"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/FactureDTO"
            }
          }
        }
      }
    },
    "/api/Facture/get-Facture-Ref/{RefFacture}": {
      "get": {
        "tags": [
          "Facture"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "RefFacture",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Facture/update-Facture/{id}": {
      "put": {
        "tags": [
          "Facture"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "factureId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FactureUpdateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Facture/delete-Facture/{id}": {
      "delete": {
        "tags": [
          "Facture"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Impaye/GetImpaye/{id}": {
      "get": {
        "tags": [
          "Impaye"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Impaye"
            }
          }
        }
      }
    },
    "/api/Impaye/CreateImpaye/{idEcheance}": {
      "post": {
        "tags": [
          "Impaye"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "idEcheance",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ImpayeCreateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        }
      }
    },
    "/api/Impaye/UpdateImpaye/{id}": {
      "put": {
        "tags": [
          "Impaye"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Impaye"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Impaye/AffecterImpayeAuDossier/{idImpaye}/{idDossier}": {
      "put": {
        "tags": [
          "Impaye"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "idImpaye",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "idDossier",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Impaye/DeleteImpaye/{id}": {
      "delete": {
        "tags": [
          "Impaye"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Abonnement": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "activationCodeReference": {
          "type": "string"
        },
        "activationDate": {
          "format": "date-time",
          "type": "string"
        },
        "expirationDate": {
          "format": "date-time",
          "type": "string"
        },
        "activationCode": {
          "type": "string"
        },
        "businessSector": {
          "type": "string"
        },
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "productReference": {
          "type": "string"
        },
        "clientRef": {
          "type": "string"
        },
        "assignedToEmail": {
          "type": "string"
        },
        "assignedToPhone": {
          "type": "string"
        },
        "isAssigned": {
          "type": "boolean",
          "readOnly": true
        },
        "maxUsersAllowed": {
          "format": "int32",
          "type": "integer"
        },
        "maxDevicesAllowed": {
          "format": "int32",
          "type": "integer"
        },
        "activationDevices": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ActivationDevice"
          }
        },
        "subscriptionType": {
          "type": "string"
        },
        "isRecurring": {
          "type": "boolean"
        },
        "isActive": {
          "type": "boolean"
        },
        "paymentMethod": {
          "type": "string"
        },
        "renewalDate": {
          "format": "date-time",
          "type": "string"
        },
        "renewalAttempts": {
          "format": "int32",
          "type": "integer"
        },
        "cancelledDate": {
          "format": "date-time",
          "type": "string"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      },
      "additionalProperties": false
    },
...
    "BonDeLivraison": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "reference": {
          "type": "string"
        },
        "titreClient": {
          "$ref": "#/definitions/Titre"
        },
        "nomClient": {
          "type": "string"
        },
        "adresseClient": {
          "type": "string"
        },
        "mfClient": {
          "type": "string"
        },
        "gsmClient": {
          "format": "int64",
          "type": "integer"
        },
        "proprietaireId": {
          "format": "int32",
          "type": "integer"
        },
        "commandesJson": {
          "type": "string"
        },
        "commandes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Commande"
          }
        },
        "montantTotalHTBL": {
          "format": "double",
          "type": "number"
        },
        "remiseBL": {
          "format": "double",
          "type": "number"
        },
        "netHT": {
          "format": "double",
          "type": "number"
        },
        "tva": {
          "format": "double",
          "type": "number"
        },
        "montantTotalTTCBL": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
...
    "Commande": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "referenceCommande": {
          "type": "string"
        },
        "nomVendeur": {
          "type": "string"
        },
        "adresseVendeur": {
          "type": "string"
        },
        "mfVendeur": {
          "type": "string"
        },
        "gsmVendeur": {
          "format": "int64",
          "type": "integer"
        },
        "proprietaireId": {
          "format": "int32",
          "type": "integer"
        },
        "titreAcheteur": {
          "$ref": "#/definitions/Titre"
        },
        "nomAcheteur": {
          "type": "string"
        },
        "adresseAcheteur": {
          "type": "string"
        },
        "mfAcheteur": {
          "type": "string"
        },
        "gsmAcheteur": {
          "format": "int64",
          "type": "integer"
        },
        "listIdsTransactions": {
          "type": "array",
          "items": {
            "format": "int32",
            "type": "integer"
          }
        },
        "montantTotalHT": {
          "format": "double",
          "type": "number"
        },
        "tva": {
          "format": "double",
          "type": "number"
        },
        "montantTotalTTC": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
...
    "Dossier": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "portefeuilleId": {
          "format": "int32",
          "type": "integer"
        },
        "portefeuille": {
          "$ref": "#/definitions/Portefeuille"
        },
        "listeImpayes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Impaye"
          }
        },
        "listeAbonnements": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Abonnement"
          }
        }
      },
      "additionalProperties": false
    },
    "Echeance": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "ref": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "dateEcheance": {
          "format": "date-time",
          "type": "string"
        },
        "montant": {
          "format": "double",
          "type": "number"
        },
        "impayeId": {
          "format": "int32",
          "type": "integer"
        },
        "impaye": {
          "$ref": "#/definitions/Impaye"
        }
      },
      "additionalProperties": false
    },
...
    "FactureCreateDTO": {
      "type": "object",
      "properties": {
        "titreClient": {
          "$ref": "#/definitions/Titre"
        },
        "nomClient": {
          "type": "string"
        },
        "adresseClient": {
          "type": "string"
        },
        "mfClient": {
          "type": "string"
        },
        "gsmClient": {
          "format": "int64",
          "type": "integer"
        },
        "proprietaireId": {
          "format": "int32",
          "type": "integer"
        },
        "listeRefsBLs": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "dateEmission": {
          "format": "date-time",
          "type": "string"
        },
        "dateMiseAJour": {
          "format": "date-time",
          "type": "string"
        },
        "tva": {
          "format": "double",
          "type": "number"
        },
        "remise": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
    "FactureDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "referenceFacture": {
          "type": "string"
        },
        "titreClient": {
          "$ref": "#/definitions/Titre"
        },
        "nomClient": {
          "type": "string"
        },
        "adresseClient": {
          "type": "string"
        },
        "mfClient": {
          "type": "string"
        },
        "gsmClient": {
          "format": "int64",
          "type": "integer"
        },
        "proprietaireId": {
          "format": "int32",
          "type": "integer"
        },
        "listeBL": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BonDeLivraison"
          }
        },
        "dateEmission": {
          "format": "date-time",
          "type": "string"
        },
        "dateMiseAJour": {
          "format": "date-time",
          "type": "string"
        },
        "montantHTFacture": {
          "format": "double",
          "type": "number"
        },
        "tva": {
          "format": "double",
          "type": "number"
        },
        "remise": {
          "format": "double",
          "type": "number"
        },
        "montantTTCFacture": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
    "FactureUpdateDTO": {
      "type": "object",
      "properties": {
        "titreClient": {
          "$ref": "#/definitions/Titre"
        },
        "nomClient": {
          "type": "string"
        },
        "adresseClient": {
          "type": "string"
        },
        "mfClient": {
          "type": "string"
        },
        "gsmClient": {
          "format": "int64",
          "type": "integer"
        },
        "listeRefsBLs": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "dateEmission": {
          "format": "date-time",
          "type": "string"
        },
        "dateMiseAJour": {
          "format": "date-time",
          "type": "string"
        },
        "tva": {
          "format": "double",
          "type": "number"
        },
        "remise": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
...
    "Impaye": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "ref": {
          "type": "string"
        },
        "refClt": {
          "type": "string"
        },
        "nom": {
          "type": "string"
        },
        "prenom": {
          "type": "string"
        },
        "createdate": {
          "type": "string"
        },
        "datetombe": {
          "type": "string"
        },
        "retard": {
          "type": "string"
        },
        "statut": {
          "type": "string"
        },
        "priorite": {
          "type": "string"
        },
        "nbreImpaye": {
          "format": "int32",
          "type": "integer"
        },
        "montant": {
          "type": "string"
        },
        "dossierId": {
          "format": "int32",
          "type": "integer"
        },
        "dossier": {
          "$ref": "#/definitions/Dossier"
        },
        "echeanceId": {
          "format": "int32",
          "type": "integer"
        },
        "echeance": {
          "$ref": "#/definitions/Echeance"
        }
      },
      "additionalProperties": false
    },
    "ImpayeCreateDTO": {
      "type": "object",
      "properties": {
        "refClt": {
          "type": "string"
        },
        "nom": {
          "type": "string"
        },
        "prenom": {
          "type": "string"
        },
        "createdate": {
          "type": "string"
        },
        "datetombe": {
          "type": "string"
        },
        "retard": {
          "type": "string"
        },
        "statut": {
          "type": "string"
        },
        "priorite": {
          "type": "string"
        },
        "montant": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "Portefeuille": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nom": {
          "type": "string"
        },
        "nbrDossiers": {
          "format": "int32",
          "type": "integer"
        },
        "date": {
          "type": "string"
        },
        "action": {
          "type": "string"
        },
        "listeDossiers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Dossier"
          }
        },
        "listeRapports": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Rapport"
          }
        },
        "userPortefeuilleAssociations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserPortefeuilleAssociation"
          }
        }
      },
      "additionalProperties": false
    },
...
    "Rapport": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "portefeuilleId": {
          "format": "int32",
          "type": "integer"
        },
        "portefeuille": {
          "$ref": "#/definitions/Portefeuille"
        }
      },
      "additionalProperties": false
    },
...
    "Titre": {
      "format": "int32",
      "enum": [
        0,
        1
      ],
      "type": "integer"
    },
...
    "UserPortefeuilleAssociation": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "portefeuilleId": {
          "format": "int32",
          "type": "integer"
        },
        "portefeuille": {
          "$ref": "#/definitions/Portefeuille"
        }
      },
      "additionalProperties": false
    },
...
  }
}